pyproject.toml
[project]
name = "x-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyboard>=0.13.5",
    "pillow>=12.1.0",
    "pyqt6>=6.10.2",
    "pywin32>=311",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
]
